pyproject.toml
[tool.poetry]
name = "pypi-files"
version = "0.1.0"
description = ""
authors = ["rcmdnk <rcmdnk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fire = "^0.4.0"
PyYAML = "^5.4.1"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
pf = "pypi_files:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
